requirements.txt
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
anthropic==0.39.0
requests==2.32.3
